+++ src/main/java/hpbui/gamerportal/repository/AccountRepository.java
@@ -3,7 +3,7 @@ package hpbui.gamerportal.repository;
 import hpbui.gamerportal.entity.Account;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface AccountRepository extends PagingAndSortingRepository<Account, Integer> {
+public interface AccountRepository extends PagingAndSortingRepository<Account, Long> {
 	Account findByEmail(String email);
 
     Account findAccountByUsername(String username);
